requirements.txt
waitress==2.0.0
discord.py==1.7.3
Werkzeug==2.0.1
steamsignin==1.0.0
requests==2.26.0
Flask_PyMongo==2.3.0
Flask==2.0.1
Flask[async]==2.0.1
pymongo==3.12.0
discord==1.7.3
